refactor: extract components into separate files
