+++ App.tsx
@@ -1,152 +1,14 @@
-
 import React, { useState, useCallback, useEffect, useReducer } from 'react';
-import type { FullTimelineData, TimelineEvent, AlternativeTimelineEvent, DivergencePoint, Language } from './types';
+import type { FullTimelineData, AlternativeTimelineEvent, DivergencePoint, Language } from './types';
 import { generateTimelineData, generateOrEditImage, generateAlternativeTimeline } from './services/geminiService';
-import { translations, type Translation } from './i18n';
+import { translations } from './i18n';
 import Spinner from './components/Spinner';
-import ImageSkeleton from './components/ImageSkeleton';
-import { ClockIcon, BranchIcon, HistoryIcon } from './components/Icons';
+import { ClockIcon, HistoryIcon } from './components/Icons';
 import esFlag from './assets/es.svg';
 import gbFlag from './assets/gb.svg';
-
-const HeroSection: React.FC<{ onGenerate: (character: string) => void; isLoading: boolean; t: Translation }> = ({ onGenerate, isLoading, t }) => {
-    const [character, setCharacter] = useState('Blas de Lezo');
-
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault();
-        if (character.trim() && !isLoading) {
-            onGenerate(character.trim());
-        }
-    };
-
-    return (
-        <div className="text-center p-8 bg-gray-900/50 rounded-lg shadow-2xl shadow-cyan-500/10 backdrop-blur-sm">
-            <h1 className="text-5xl font-bold text-transparent bg-clip-text bg-gradient-to-r from-cyan-400 to-teal-200 mb-4">{t.appTitle}</h1>
-            <p className="text-lg text-gray-300 mb-6">{t.appDescription}</p>
-
-            <div className="text-left max-w-2xl mx-auto text-gray-300 mb-8">
-                <p className="mb-3 font-semibold">{t.stepsIntro}</p>
-                <ol className="list-decimal list-inside space-y-2">
-                    <li className="bg-gray-800/40 border-l-4 border-cyan-500 p-3">
-                        <div className="flex items-start gap-2">
-                            <ClockIcon className="w-5 h-5 text-cyan-400 mt-1" />
-                            <span>{t.step1}</span>
-                        </div>
-                    </li>
-                    <li className="bg-gray-800/40 border-l-4 border-cyan-500 p-3">
-                        <div className="flex items-start gap-2">
-                            <BranchIcon className="w-5 h-5 text-cyan-400 mt-1" />
-                            <span>{t.step2}</span>
-                        </div>
-                    </li>
-                    <li className="bg-gray-800/40 border-l-4 border-cyan-500 p-3">
-                        <div className="flex items-start gap-2">
-                            <HistoryIcon className="w-5 h-5 text-cyan-400 mt-1" />
-                            <span>{t.step3}</span>
-                        </div>
-                    </li>
-                </ol>
-            </div>
-
-            <form onSubmit={handleSubmit} className="flex flex-col sm:flex-row justify-center items-center gap-4">
-                <input
-                    type="text"
-                    value={character}
-                    onChange={(e) => setCharacter(e.target.value)}
-                    placeholder={t.placeholder}
-                    aria-label={t.placeholder}
-                    className="w-full sm:w-80 px-4 py-3 bg-gray-800/70 border-2 border-gray-600 rounded-lg text-white focus:outline-none focus:ring-2 focus:ring-cyan-400 focus:border-cyan-400 transition duration-300"
-                    disabled={isLoading}
-                />
-                <button
-                    type="submit"
-                    className="w-full sm:w-auto px-8 py-3 bg-gradient-to-r from-cyan-500 to-teal-500 text-white font-semibold rounded-lg shadow-lg hover:shadow-cyan-500/50 transform hover:scale-105 transition duration-300 disabled:opacity-50 disabled:cursor-not-allowed"
-                    disabled={isLoading}
-                >
-                    {isLoading ? <Spinner size="sm" label={t.loading} /> : t.generateButton}
-                </button>
-            </form>
-        </div>
-    );
-};
-
-const ImageCard: React.FC<{ eventId: number; images: Record<string, string>; imageLoading: Set<number>; alt: string; errorText: string; loadingLabel: string; }> = ({ eventId, images, imageLoading, alt, errorText, loadingLabel }) => {
-    return (
-        <div className="relative aspect-video bg-gray-800/50 rounded-lg flex items-center justify-center overflow-hidden border border-gray-700 shadow-lg">
-            {imageLoading.has(eventId) && <ImageSkeleton />}
-            {imageLoading.has(eventId) && <Spinner size="md" label={loadingLabel} />}
-            {!imageLoading.has(eventId) && !images[eventId] && <p className="text-gray-500">{errorText}</p>}
-            {images[eventId] && (
-                <img
-                    src={images[eventId]}
-                    alt={alt}
-                    className="w-full h-full object-cover transition-opacity duration-500 opacity-0"
-                    onLoad={(e) => (e.currentTarget.style.opacity = '1')}
-                />
-            )}
-        </div>
-    );
-};
-
-const TimelineNode: React.FC<{ isLast?: boolean }> = ({ isLast = false }) => (
-    <div className="absolute left-1/2 -ml-[1px] top-12 w-0.5 h-full bg-gray-700">
-        <div className="absolute left-1/2 -ml-2 top-0 w-4 h-4 rounded-full bg-gray-700 border-2 border-cyan-400"></div>
-        {isLast && <div className="absolute left-1/2 -ml-2 bottom-0 w-4 h-4 rounded-full bg-gray-700 border-2 border-cyan-400"></div>}
-    </div>
-);
-
-// Fix: Omit derived props from ImageCard as they are provided within this component.
-const EventCard: React.FC<{
-    event: TimelineEvent | AlternativeTimelineEvent;
-    divergence?: DivergencePoint;
-    onShowAlternative: (divergence: DivergencePoint) => void;
-    isAlternative?: boolean;
-    t: Translation;
-} &
-    Omit<
-        React.ComponentProps<typeof ImageCard>,
-        'eventId' | 'errorText' | 'alt' | 'loadingLabel'
-    > &
-    React.HTMLAttributes<HTMLDivElement>> = ({
-    event,
-    divergence,
-    onShowAlternative,
-    isAlternative = false,
-    t,
-    className = '',
-    style,
-    ...imageProps
-}) => {
-    const title = 'titulo' in event ? event.titulo : event.titulo_eco;
-
-    return (
-        <div
-            className={`relative group w-full lg:w-2/3 mx-auto p-4 ${className}`}
-            style={style}
-        >
-            <div className="bg-gray-800/30 backdrop-blur-md rounded-xl shadow-lg border border-gray-700/50 p-6 transition-all duration-300 hover:border-cyan-400/50 hover:shadow-cyan-500/10">
-                <h3 className="text-2xl font-bold text-cyan-300 mb-2">{title}</h3>
-                <p className="text-gray-300 mb-4">{event.descripcion_corta}</p>
-                <ImageCard
-                    eventId={event.id}
-                    alt={t.imageAlt(title)}
-                    errorText={t.imageError}
-                    loadingLabel={t.loading}
-                    {...imageProps}
-                />
-                {divergence && !isAlternative && (
-                    <div className="mt-6 p-4 border-t-2 border-dashed border-gray-600">
-                        <h4 className="text-lg font-semibold text-yellow-300 mb-3 flex items-center"><BranchIcon className="w-5 h-5 mr-2"/> {t.divergenceTitle}</h4>
-                        <p className="italic text-gray-400 mb-4">"{divergence.titulo_pregunta}"</p>
-                        <button onClick={() => onShowAlternative(divergence)} className="px-4 py-2 bg-yellow-500/20 text-yellow-300 border border-yellow-400 rounded-lg hover:bg-yellow-500/40 transition-colors">
-                            {t.exploreAlternative}
-                        </button>
-                    </div>
-                )}
-            </div>
-        </div>
-    );
-};
+import HeroSection from './components/HeroSection';
+import EventCard from './components/EventCard';
+import TimelineNode from './components/TimelineNode';
 
 type AppState = {
     characterName: string;
